+++ src/SplitTree.java
@@ -3,12 +3,12 @@ public class SplitTree extends AbstractTree{
 	
 
 	
-	public SplitTree(Turtle t, int level, int length, float angle, int thickness, int color) {
+	public SplitTree(Turtle t, int level, float length, float angle, int thickness, int color) {
 		super(t, level, length, angle, thickness, color);
 	}
 
 	@Override
-	public void drawTree(Turtle t, int level, int length, float angle, int thickness, int color) {
+	public void drawTree(Turtle t, int level, float length, float angle, int thickness, int color) {
 		
 	}
 }
